requirements.txt
# ─── Core UI ──────────────────────────────────────────────────────────────────
streamlit==1.44.0

# ─── Ollama / OpenAI-compatible client ───────────────────────────────────────
openai==1.68.2

# ─── PDF → image conversion (PyMuPDF — import name: fitz) ────────────────────
pymupdf==1.25.3

# ─── Image processing ────────────────────────────────────────────────────────
Pillow==11.1.0

# ─── Data manipulation and Excel export ──────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5          # pandas xlsx engine

# ─── PDF report generation ───────────────────────────────────────────────────
reportlab==4.2.5

# ─── Environment variable loading (local dev) ────────────────────────────────
python-dotenv==1.0.1

# ─── Tesseract OCR fallback (system tesseract-ocr must also be installed) ────
pytesseract==0.3.13
